adds Top-bar Component
