+++ src/components/primitive/index.ts
@@ -14,6 +14,8 @@ export * from './icon-svg/icon-svg.component.props';
 
 export * from './image/image.component';
 
+export * from './avatar/avatar.component';
+
 export * from './navlink/navlink.component';
 export * from './navlink/navlink.component.props';
